move `despacito.js` into `silly` folder + category, split videos into separate file
and just some small changes. nothing insane. for now.
and also removed cooldown (uses default).
